Rewrite the install.php
I startet to rewrite the install.php so it's based on classes
instead of the "spagetti-code" witch functions and no
structure/logic behind it.

I used the rewrite to add some new features:
  * The shell now supports wget and curl.
  * The user can choose witch version should be installed. The tags
    are dynamicly fetched from the github api, so there is no need
    to update the contao-check on every new release.
  * The user can specify in witch directory the new contao should
    be installed.

The html part is currently not complete rewritten, but sience i hate
css thats a job for some of the frontend developers here. I am
focusing on rewriting the other scripts.